requirements.txt
Flask
slackclient

